feat(web): align the public law shell with the workspace chrome
The public shell now renders the same h-12 top bar (breadcrumb with the
current decision), the shared sidebar footer (feedback entry plus the
user menu extracted into SidebarUserMenu), and an anonymous twin of the
inspector side rail whose affordances open the sign-in dialog.
